Move temporary certificate policy init into policy package
