+++ test/e2e/suite/secrettemplate/BUILD.bazel
@@ -6,7 +6,6 @@ go_library(
     importpath = "github.com/jetstack/cert-manager/test/e2e/suite/secrettemplate",
     visibility = ["//visibility:public"],
     deps = [
-        "//internal/controller/feature:go_default_library",
         "//pkg/apis/certmanager/v1:go_default_library",
         "//pkg/apis/meta/v1:go_default_library",
         "//test/e2e/framework:go_default_library",
